Extract reachability streaming to reachability.py
Move the bodies of subscribe_reachability, _reachability_refresh_loop,
_build_reachability_snapshot, _on_reachability_observation,
get_reachability_snapshot, and refresh_device_mdns into a sibling
module. The controller keeps thin bound-method delegates so the WS
dispatch, the ReachabilityTracker callback, and tests that call
these as instance methods all resolve to the same names.

Pure refactor; behaviour preserved. Drops ~135 lines from
controllers/devices/controller.py.
